refactor(rendering): drop unread context fields, consolidate test fixture, refresh module doc
Three small cleanups in the resolution module:

- ResolutionContext carried output_intent_cmyk and rendering_intent
  fields that no resolver stage read. Every caller dutifully forwarded
  them anyway. Drop them — they will be added back when the colour
  stage grows an ICC code path that actually consumes them.

- The same fixture_doc PDF-builder was copy-pasted into three test
  modules (color.rs, pipeline.rs, context.rs). Extract into a shared
  test_support module gated behind cfg(test).

- The mod.rs doc-comment still described the pipeline as 'dead at
  landing' with one toggled pilot operator, which is stale: f/f*/S/s/
  B/B*/b/b* all route through the pipeline behind the env-var toggle
  today. Rewrite the status paragraph and the dead-code allow comment
  to reflect the actual scope and reason.
